Refactor BlockMeta
