+++ core/server/worker/src/main/java/alluxio/worker/block/meta/TempBlockMeta.java
@@ -11,15 +11,47 @@
 
 package alluxio.worker.block.meta;
 
+import alluxio.conf.PropertyKey;
+import alluxio.conf.ServerConfiguration;
+import alluxio.util.io.PathUtils;
+import alluxio.worker.block.BlockStoreLocation;
+
+import com.google.common.base.Preconditions;
+
 import javax.annotation.concurrent.NotThreadSafe;
 
 /**
  * Represents the metadata of an uncommitted block in Alluxio managed storage.
  */
 @NotThreadSafe
-public final class TempBlockMeta extends AbstractBlockMeta {
+public final class TempBlockMeta {
+  private final long mBlockId;
+  private final StorageDir mDir;
   private final long mSessionId;
   private long mTempBlockSize;
+
+  /**
+   * All blocks are created as temp blocks before committed. They are stored in BlockStore under a
+   * subdir of its {@link StorageDir}, the subdir is tmpFolder/sessionId % maxSubdirMax.
+   * tmpFolder is a property of {@link PropertyKey#WORKER_DATA_TMP_FOLDER}.
+   * maxSubdirMax is a property of {@link PropertyKey#WORKER_DATA_TMP_SUBDIR_MAX}.
+   * The block file name is "sessionId-blockId". e.g. sessionId 2 creates a temp Block 100 in
+   * {@link StorageDir} "/mnt/mem/0", this temp block has path:
+   * <p>
+   * /mnt/mem/0/.tmp_blocks/2/2-100
+   *
+   * @param dir the parent directory
+   * @param sessionId the session id
+   * @param blockId the block id
+   * @return temp file path
+   */
+  public static String tempPath(StorageDir dir, long sessionId, long blockId) {
+    final String tmpDir = ServerConfiguration.get(PropertyKey.WORKER_DATA_TMP_FOLDER);
+    final int subDirMax = ServerConfiguration.getInt(PropertyKey.WORKER_DATA_TMP_SUBDIR_MAX);
+
+    return PathUtils.concatPath(dir.getDirPath(), tmpDir, sessionId % subDirMax,
+        String.format("%x-%x", sessionId, blockId));
+  }
 
   /**
    * Creates a new instance of {@link TempBlockMeta}.
@@ -30,26 +62,53 @@ public final class TempBlockMeta extends AbstractBlockMeta {
    * @param dir {@link StorageDir} of this temp block belonging to
    */
   public TempBlockMeta(long sessionId, long blockId, long initialBlockSize, StorageDir dir) {
-    super(blockId, dir);
+    mBlockId = blockId;
+    mDir = Preconditions.checkNotNull(dir, "dir");
     mSessionId = sessionId;
     mTempBlockSize = initialBlockSize;
   }
 
-  @Override
+  /**
+   * @return the block size
+   */
   public long getBlockSize() {
     return mTempBlockSize;
   }
 
-  @Override
+  /**
+   * @return the block path
+   */
   public String getPath() {
     return tempPath(mDir, mSessionId, mBlockId);
+  }
+
+  /**
+   * @return the block id
+   */
+  public long getBlockId() {
+    return mBlockId;
+  }
+
+  /**
+   * @return location of the block
+   */
+  public BlockStoreLocation getBlockLocation() {
+    StorageTier tier = mDir.getParentTier();
+    return new BlockStoreLocation(tier.getTierAlias(), mDir.getDirIndex(), mDir.getDirMedium());
+  }
+
+  /**
+   * @return the parent directory
+   */
+  public StorageDir getParentDir() {
+    return mDir;
   }
 
   /**
    * @return the commit path
    */
   public String getCommitPath() {
-    return commitPath(mDir, mBlockId);
+    return BlockMeta.commitPath(mDir, mBlockId);
   }
 
   /**
